+++ router.go
@@ -17,4 +17,5 @@ func customizeRegister(r *server.Hertz) {
 	r.GET("/v1/tongji/oauth/authorize", handler.TongjiAuthorize)
 	r.POST("/v1/tongji/oauth/token", handler.TongjiExchangeToken)
 	r.OPTIONS("/v1/tongji/oauth/token", handler.TongjiExchangeTokenOptions)
+	r.GET("/v1/tongji/user/basic-info", handler.TongjiUserBasicInfo)
 }